Moves a test to a better place
